Non functional non-trivial authorization

+++ app/Http/Livewire/Gradings.php
@@ -7,18 +7,30 @@ use App\Models\Grading;
 use Livewire\Component;
 use Illuminate\Validation\Rule;
 use Illuminate\Support\Collection;
+use Illuminate\Support\Facades\App;
 use Illuminate\Support\Facades\Log;
+use Illuminate\Auth\Access\AuthorizationException;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class Gradings extends Component
 {
+    use AuthorizesRequests;
 
     public $gradingId;
+
+    public $trashed = false;
 
     public $name;
     public $values = [];
 
-    public function mount()
+    /**
+     * Component lifecyle method that executes once, when the component is mounting
+     * 
+     * @param string $trashed
+     */
+    public function mount(string $trashed = null)
     {
+        $this->trashed = boolval($trashed);
         /** @var Collection */
         $grades = Grade::all(['grade', 'points']);
 
@@ -33,6 +45,11 @@ class Gradings extends Component
         
     }
 
+    /**
+     * Lifecycle method that renders the component evrytime the stte of the componen changes
+     * 
+     * @return View
+     */
     public function render()
     {
         return view('livewire.gradings', [
@@ -41,11 +58,23 @@ class Gradings extends Component
         ]);
     }
 
+    /**
+     * Get all grading systems from the database
+     */
     public function getAllGradings()
     {
-        return Grading::all();
+        $gradingsQuery = Grading::query();
+
+        if($this->trashed) $gradingsQuery->onlyTrashed();
+
+        return $gradingsQuery->get();
     }
 
+    /**
+     * Compoent models properties validtion fields
+     * 
+     * @return array
+     */
     public function rules()
     {
         return [
@@ -58,11 +87,18 @@ class Gradings extends Component
         ];
     }
 
+    /**
+     * Adding a grading system enrty to the database
+     */
     public function addGrading()
     {
         $data = $this->validate();
 
         try {
+
+            $this->authorize('create', Grading::class);
+
+            /** @var Grading */
             $grading = Grading::create($data);
 
             if($grading){
@@ -78,11 +114,9 @@ class Gradings extends Component
 
         } catch (\Exception $exception) {
 
-            Log::error($exception->getMessage(), [
-                'action' => __METHOD__
-            ]);
-
-            session()->flash('error', 'A fatal error occurred');
+            Log::error($exception->getMessage(), ['action' => __METHOD__]);
+            
+            $this->setError($exception, "Sorry, adding grading system operation failed");
 
             $this->emit('hide-upsert-grading-modal');
             
@@ -90,6 +124,26 @@ class Gradings extends Component
         
     }
 
+    /**
+     * Set appropriate error based on the type of the error and the environment
+     * 
+     * @param \Exception $exception
+     * @param string $message
+     */
+    private function setError(\Exception $exception, string $message)
+    {
+        if($exception instanceof AuthorizationException) $message = $exception->getMessage();
+
+        else $message = App::environment('local') ? $exception->getMessage() : $message;
+
+        session()->flash('error', $message);
+    }  
+
+    /**
+     * Show upsert grading model for updating the specified grading
+     * 
+     * @param Grading $grading
+     */
     public function editGrading(Grading $grading)
     {
         $this->gradingId = $grading->id;
@@ -100,6 +154,9 @@ class Gradings extends Component
         $this->emit('show-upsert-grading-modal');
     }
 
+    /**
+     * Update a grading database entry
+     */
     public function updateGrading()
     {
         $data = $this->validate();
@@ -108,6 +165,8 @@ class Gradings extends Component
 
             /** @var Grading */
             $grading = Grading::findOrFail($this->gradingId);
+
+            $this->authorize('update', $grading);
 
             if ($grading->update($data)) {
 
@@ -123,17 +182,20 @@ class Gradings extends Component
             
         } catch (\Exception $exception) {
 
-            Log::error($exception->getMessage(), [
-                'action' => __METHOD__
-            ]);
-
-            session()->flash('error', 'A fatal error occurred');
+            Log::error($exception->getMessage(), ['action' => __METHOD__]);
+            
+            $this->setError($exception, "Sorry, updating grading system operation failed");
 
             $this->emit('hide-upsert-grading-modal');
             
         }
     }
 
+    /**
+     * Show confirmation modal for deleting a grading system
+     * 
+     * @param Grading $grading
+     */
     public function showDeleteGradingModal(Grading $grading)
     {
         $this->gradingId = $grading->id;
@@ -143,6 +205,9 @@ class Gradings extends Component
         $this->emit('show-delete-grading-modal');
     }
 
+    /**
+     * Trashing a grading system
+     */
     public function deleteGrading()
     {
 
@@ -150,6 +215,8 @@ class Gradings extends Component
 
             /** @var Grading */
             $grading = Grading::findOrFail($this->gradingId);
+
+            $this->authorize('delete', $grading);
 
             if ($grading->delete()) {
 
@@ -165,11 +232,9 @@ class Gradings extends Component
             
         } catch (\Exception $exception) {
 
-            Log::error($exception->getMessage(), [
-                'action' => __METHOD__
-            ]);
-
-            session()->flash('error', 'A fatal error occurred');
+            Log::error($exception->getMessage(), ['action' => __METHOD__]);
+            
+            $this->setError($exception, "Sorry, failed deleting a grading system");
 
             $this->emit('hide-delete-grading-modal');
             
@@ -177,6 +242,11 @@ class Gradings extends Component
         
     }
 
+    /**
+     * Show the details of a grading system
+     * 
+     * @param Grading $grading
+     */
     public function showGrading(Grading $grading)
     {
         $this->name = $grading->name;
@@ -184,5 +254,58 @@ class Gradings extends Component
 
         $this->emit('show-grading-instance-modal');
         
+    }
+
+    /**
+     * Restore a deleted grading system
+     * 
+     * @param mixed $gradingId
+     */
+    public function restoreGrading($gradingId)
+    {
+        try {
+            /** @var Grading */
+            $grading = Grading::where('id', $gradingId)->withTrashed()->firstOrFail();
+
+            $this->authorize('restore', $grading);
+
+            $grading->restore();
+
+            session()->flash('status', "The grading system, {$grading->name}, has been restored");
+
+        } catch (\Exception $exception) {
+
+            Log::error($exception->getMessage(), ['action' => __METHOD__]);
+            
+            $this->setError($exception, "Sorry, failed restoring the grading system");
+            
+        }
+    }
+
+    /**
+     * Completely deleted a grading system from the database
+     * 
+     * @param mixed $gradingId
+     */
+    public function destroyGrading($gradingId)
+    {
+        try {
+
+            /** @var Grading */
+            $grading = Grading::where('id', $gradingId)->withTrashed()->firstOrFail();
+
+            $this->authorize('forceDelete', $grading);
+
+            $grading->forceDelete();
+
+            session()->flash('status', "The grading system, {$grading->name}, has been completely deleted");
+
+        } catch (\Exception $exception) {
+
+            Log::error($exception->getMessage(), ['action' => __METHOD__]);
+            
+            $this->setError($exception, "Sorry, failed completely deleting the grading system");
+            
+        }
     }
 }
